+++ test/Audit.UnitTest/ConfigurationTests.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Audit.Core;
 using Audit.Core.Providers;
 using NUnit.Framework;
@@ -12,14 +13,13 @@ namespace Audit.UnitTest
         public void SetUp()
         {
             Core.Configuration.Reset();
+            Core.Configuration.DataProvider = new InMemoryDataProvider();
+            Core.Configuration.CreationPolicy = EventCreationPolicy.Manual;
         }
 
         [Test]
         public void TestConfiguration_DataProviderAs()
         {
-            // Arrange
-            Core.Configuration.DataProvider = new InMemoryDataProvider();
-
             // Act
             var dataProvider = Core.Configuration.DataProviderAs<InMemoryDataProvider>();
 
@@ -33,8 +33,6 @@ namespace Audit.UnitTest
         {
             // Arrange
             int test = 0;
-            Core.Configuration.DataProvider = new InMemoryDataProvider();
-            Core.Configuration.CreationPolicy = EventCreationPolicy.Manual;
 
             // Act
             Core.Configuration.AddCustomAction(ActionType.OnEventSaved, async (scope, ct) =>
@@ -54,14 +52,9 @@ namespace Audit.UnitTest
         {
             // Arrange
             int test = 0;
-            Core.Configuration.DataProvider = new InMemoryDataProvider();
-            Core.Configuration.CreationPolicy = EventCreationPolicy.Manual;
-            
+
             // Act
-            Core.Configuration.AddOnSavingAction(scope =>
-            {
-                test++;
-            });
+            Core.Configuration.AddOnSavingAction(scope => { test++; });
 
             var scope = AuditScope.Create("test", null);
             scope.Save();
@@ -75,8 +68,6 @@ namespace Audit.UnitTest
         {
             // Arrange
             int test = 0;
-            Core.Configuration.DataProvider = new InMemoryDataProvider();
-            Core.Configuration.CreationPolicy = EventCreationPolicy.Manual;
 
             // Act
             Core.Configuration.AddOnSavingAction(async (scope, ct) =>
@@ -97,8 +88,6 @@ namespace Audit.UnitTest
         {
             // Arrange
             int test = 0;
-            Core.Configuration.DataProvider = new InMemoryDataProvider();
-            Core.Configuration.CreationPolicy = EventCreationPolicy.Manual;
 
             // Act
             Core.Configuration.AddOnSavingAction(async scope =>
@@ -119,8 +108,6 @@ namespace Audit.UnitTest
         {
             // Arrange
             int test = 0;
-            Core.Configuration.DataProvider = new InMemoryDataProvider();
-            Core.Configuration.CreationPolicy = EventCreationPolicy.Manual;
 
             // Act
             Core.Configuration.AddOnCreatedAction(async (scope, ct) =>
@@ -141,8 +128,6 @@ namespace Audit.UnitTest
         {
             // Arrange
             int test = 0;
-            Core.Configuration.DataProvider = new InMemoryDataProvider();
-            Core.Configuration.CreationPolicy = EventCreationPolicy.Manual;
 
             // Act
             Core.Configuration.AddOnCreatedAction(async scope =>
@@ -165,8 +150,6 @@ namespace Audit.UnitTest
         public void TestConfiguration_ResetCustomActionsByType(ActionType type)
         {
             // Arrange
-            Core.Configuration.DataProvider = new InMemoryDataProvider();
-            Core.Configuration.CreationPolicy = EventCreationPolicy.Manual;
             Core.Configuration.AddCustomAction(ActionType.OnScopeCreated, _ => { });
             Core.Configuration.AddCustomAction(ActionType.OnEventSaving, _ => { });
             Core.Configuration.AddCustomAction(ActionType.OnEventSaved, _ => { });
@@ -184,14 +167,10 @@ namespace Audit.UnitTest
         public void TestConfiguration_AddOnDisposedAction()
         {
             int onDisposed = 0;
-            Core.Configuration.AddOnDisposedAction(s =>
-            {
-                onDisposed++; 
-
-            });
+            Core.Configuration.AddOnDisposedAction(s => { onDisposed++; });
             Core.Configuration.AddOnDisposedAction(async s =>
             {
-                await Task.Yield(); 
+                await Task.Yield();
                 onDisposed++;
             });
             Core.Configuration.AddOnDisposedAction(async (s, ct) =>
@@ -207,8 +186,166 @@ namespace Audit.UnitTest
 
             var scope = AuditScope.Create(new AuditScopeOptions() { DataProvider = new NullDataProvider() });
             scope.Dispose();
-            
+
             Assert.That(onDisposed, Is.EqualTo(4));
+        }
+
+        [TestCase(ActionType.OnScopeCreated)]
+        [TestCase(ActionType.OnEventSaving)]
+        [TestCase(ActionType.OnScopeDisposed)]
+        public void TestConfiguration_CustomActionsContinuation(ActionType type)
+        {
+            // Arrange
+            int count = 0;
+
+            // To cover the methods per action type
+            switch (type)
+            {
+                case ActionType.OnScopeCreated:
+                    Core.Configuration.AddOnCreatedAction(scope =>
+                    {
+                        count++;
+                        return true;
+                    });
+                    break;
+                case ActionType.OnEventSaving:
+                    Core.Configuration.AddOnSavingAction(scope =>
+                    {
+                        count++;
+                        return true;
+                    });
+                    break;
+                case ActionType.OnScopeDisposed:
+                    Core.Configuration.AddOnDisposedAction(scope =>
+                    {
+                        count++;
+                        return true;
+                    });
+                    break;
+                default:
+                    Core.Configuration.AddCustomAction(type, scope =>
+                    {
+                        count++;
+                        return true;
+                    });
+                    break;
+            }
+
+            Core.Configuration.AddCustomAction(type, async (scope, ct) =>
+            {
+                count++;
+                return await Task.FromResult(false);
+            });
+            Core.Configuration.AddCustomAction(type, async (scope, ct) =>
+            {
+                count = -1;
+                return await Task.FromResult(true);
+            });
+
+            // Act
+            var scope = AuditScope.Create("test", null);
+            scope.Save();
+            scope.Dispose();
+
+            // Assert
+            Assert.That(count, Is.EqualTo(2));
+        }
+
+        [TestCase(ActionType.OnScopeCreated)]
+        [TestCase(ActionType.OnEventSaving)]
+        [TestCase(ActionType.OnEventSaved)]
+        [TestCase(ActionType.OnScopeDisposed)]
+        public void TestConfiguration_CustomActionsContinuation_Async(ActionType type)
+        {
+            // Arrange
+            int count = 0;
+            Core.Configuration.AddCustomAction(type, async (scope, ct) =>
+            {
+                count = 1;
+                return await Task.FromResult(true);
+            });
+
+            // To cover the methods per action type
+            switch (type)
+            {
+                case ActionType.OnScopeCreated:
+                    Core.Configuration.AddOnCreatedAction(async (scope, ct) =>
+                    {
+                        count++;
+                        return await Task.FromResult(false);
+                    });
+                    break;
+                case ActionType.OnEventSaving:
+                    Core.Configuration.AddOnSavingAction(async (scope, ct) =>
+                    {
+                        count++;
+                        return await Task.FromResult(false);
+                    });
+                    break;
+                case ActionType.OnScopeDisposed:
+                    Core.Configuration.AddOnDisposedAction(async (scope, ct) =>
+                    {
+                        count++;
+                        return await Task.FromResult(false);
+                    });
+                    break;
+                default:
+                    Core.Configuration.AddCustomAction(type, async (scope, ct) =>
+                    {
+                        count++;
+                        return await Task.FromResult(false);
+                    });
+                    break;
+            }
+
+            Core.Configuration.AddCustomAction(type, async (scope, ct) =>
+            {
+                count++;
+                return await Task.FromResult(false);
+            });
+            Core.Configuration.AddCustomAction(type, async (scope, ct) =>
+            {
+                count = -1;
+                return await Task.FromResult(true);
+            });
+
+            // Act
+            var scope = AuditScope.Create("test", null);
+            scope.Save();
+            scope.Dispose();
+
+            // Assert
+            Assert.That(count, Is.EqualTo(2));
+        }
+
+        [TestCase(true)]
+        [TestCase(false)]
+        public async Task TestConfiguration_ExcludeEnvironmentInfo(bool exclude)
+        {
+            // Arrange
+            Core.Configuration.ExcludeEnvironmentInfo = exclude;
+
+            // Act
+            var scope = await AuditScope.CreateAsync("test", null);
+            await scope.SaveAsync();
+
+            // Assert
+            Assert.That(scope.Event.Environment, exclude ? Is.Null : Is.Not.Null);
+        }
+
+        [TestCase(true)]
+        [TestCase(false)]
+        public async Task TestConfiguration_ExcludeEnvironmentInfo_Option(bool exclude)
+        {
+            // Arrange
+            Core.Configuration.ExcludeEnvironmentInfo = !exclude;
+
+            // Act
+            var scope = await AuditScope.CreateAsync(c => c.ExcludeEnvironmentInfo(exclude));
+            await scope.SaveAsync();
+
+            // Assert
+            Assert.That(scope.Event.Environment, exclude ? Is.Null : Is.Not.Null);
         }
     }
 }